+++ ssc/ddoi_configurations/ssc_inst_config.ini
@@ -1,0 +1,1 @@
+[ob_keys]
