+++ ssc/imager/SetFrameNumber.py
@@ -1,38 +1,38 @@
 import time
 import ktl
 
-from ddoitranslatormodule.SSCTranslatorFunction import SSCTranslatorFunction
+from SSCTranslatorFunction import SSCTranslatorFunction
 from .. import (log, SSCException, FailedPreCondition, FailedPostCondition,
                 FailedToReachDestination, check_input)
 
 
-class SetImagePath(SSCTranslatorFunction):
-    '''Sets the image save path for the SSC in the magiq keyword service.
+class SetFrameNumber(SSCTranslatorFunction):
+    '''Sets the image frame number for the SSC in the magiq keyword service.
     
     Args:
-    Image Path - The path for saving images
+    frame - The frame number for saving images
     '''
     @classmethod
     def pre_condition(cls, args, logger, cfg):
-        check_input(args, 'path', allowed_types=[str])
+        check_input(args, 'frame', allowed_types=[int])
         return True
 
     @classmethod
     def perform(cls, args, logger, cfg):
         magiq = ktl.cache('magiq')
-        path = args.get('path')
-        log.debug("Setting exposure time to "+path)
-        magiq['IMGDIR'].write(path)
+        frame = args.get('frame')
+        log.debug("Setting frame number to "+frame)
+        magiq['IMGFRNR'].write(frame)
         magiq['IMGCMD'].write('set')
 
     @classmethod
     def post_condition(cls, args, logger, cfg):
         log.debug("Checking for success")
-        path = args.get('path')
+        frame = args.get('frame')
         magiq = ktl.cache('magiq')
-        magiqpath=magiq.read('IMGDIR')
-        if magiqpath!=path:
-            raise FailedToReachDestination(magiqpath, path)
+        magiqframe=magiq.read('IMGFRNR')
+        if magiqframe!=frame:
+            raise FailedToReachDestination(magiqframe, frame)
         else:
             return True
 
@@ -42,7 +42,7 @@ class SetImagePath(SSCTranslatorFunction):
         '''
         from collections import OrderedDict
         args_to_add = OrderedDict()
-        args_to_add['path'] = {'type': str,
-                                  'help': 'The path for saving images.'}
+        args_to_add['frame'] = {'type': int,
+                                  'help': 'The frame number for saving images.'}
         parser = cls._add_args(parser, args_to_add, print_only=False)
         return super().add_cmdline_args(parser, cfg)
