+++ ssc/imager/SetImageSave.py
@@ -1,41 +1,43 @@
 import time
 import ktl
 
-from ddoitranslatormodule.SSCTranslatorFunction import SSCTranslatorFunction
+from SSCTranslatorFunction import SSCTranslatorFunction
 from .. import (log, SSCException, FailedPreCondition, FailedPostCondition,
                 FailedToReachDestination, check_input)
 
 
-class SetExptime(SSCTranslatorFunction):
-    '''Sets the exposure time for the SSC in the magiq keyword service.
+class SetImageSave(SSCTranslatorFunction):
+    '''Sets the image saving parameter for the SSC in the magiq keyword service.
     
     Args:
-    Exptime - The exposure time in seconds
+    Image Save - Turns image saving on/off
     '''
     @classmethod
     def pre_condition(cls, args, logger, cfg):
-        check_input(args, 'Exptime', allowed_types=[int, float])
+        check_input(args, 'save', allowed_types=[bool])
         return True
 
     @classmethod
     def perform(cls, args, logger, cfg):
         magiq = ktl.cache('magiq')
-        ttime = args.get('TTIME')
-        log.debug(f"Setting exposure time to {ttime:.3f}")
-        magiq['TTIME'].write(ttime)
-        magiq['camcmd'].write('start')
+        save = args.get('save')
+        log.debug("Setting image saving to "+str(save))
+        if save==True:
+            magiq['mqsnpff'].write(1)
+        else:
+            magiq['mqsnpff'].write(0)
+
 
     @classmethod
     def post_condition(cls, args, logger, cfg):
         log.debug("Checking for success")
-        ttime = args.get('TTIME')
+        save = args.get('save')
         magiq = ktl.cache('magiq')
-        magiqttime = magiq.read('TTIME')
-        if magiqttime!=ttime:
-            raise FailedToReachDestination(magiqttime, ttime)
+        magiqsave=magiq.read('mqsnpff')
+        if save!=bool(magiqsave):
+            raise FailedToReachDestination(magiqsave, save)
         else:
             return True
-
 
     @classmethod
     def add_cmdline_args(cls, parser, cfg=None):
@@ -43,7 +45,7 @@ class SetExptime(SSCTranslatorFunction):
         '''
         from collections import OrderedDict
         args_to_add = OrderedDict()
-        args_to_add['Exptime'] = {'type': float,
-                                  'help': 'The exposure time in seconds.'}
+        args_to_add['save'] = {'type': bool,
+                                  'help': 'Set image saving to True/False.'}
         parser = cls._add_args(parser, args_to_add, print_only=False)
         return super().add_cmdline_args(parser, cfg)
